report: extract independent report types
