+++ clients/.eslintrc.js
@@ -6,7 +6,6 @@
 'use strict';
 
 module.exports = {
-  extends: '../.eslintrc.js',
   env: {
     browser: true,
   },
