+++ lighthouse-viewer/types/viewer.d.ts
@@ -5,16 +5,13 @@
  */
 
 import _ReportGenerator = require('../../report/generator/report-generator.js');
-import {DOM as _DOM} from '../../report/renderer/dom.js';
-import {ReportRenderer as _ReportRenderer} from '../../report/renderer/report-renderer.js';
-import {ReportUIFeatures as _ReportUIFeatures} from '../../report/renderer/report-ui-features.js';
 import {Logger as _Logger} from '../../report/renderer/logger.js';
-import {TextEncoding as _TextEncoding} from '../../report/renderer/text-encoding.js';
 import {LighthouseReportViewer as _LighthouseReportViewer} from '../app/src/lighthouse-report-viewer.js';
-import 'google.analytics';
 import {FirebaseNamespace} from '@firebase/app-types';
-import '@firebase/auth-types';
+import _ReportResult from '../../report/types/report-result';
 
+// Import for needed DOM type augmentation.
+import '../../report/types/augment-dom';
 
 declare global {
   var ReportGenerator: typeof _ReportGenerator;
@@ -29,7 +26,9 @@ declare global {
     // Inserted by viewer build.
     LH_CURRENT_VERSION: string;
   }
+  
+  // Expose global types in LH namespace.
+  module LH {
+    export import ReportResult = _ReportResult;
+  }
 }
-
-// empty export to keep file a module
-export {}
