+++ report/renderer/util.js
@@ -388,7 +388,7 @@ export class Util {
   }
 
   /**
-   * @param {LH.Config.Settings} settings
+   * @param {LH.Result['configSettings']} settings
    * @return {!Array<{name: string, description: string}>}
    */
   static getEnvironmentDisplayValues(settings) {
@@ -411,7 +411,7 @@ export class Util {
   }
 
   /**
-   * @param {LH.Config.Settings} settings
+   * @param {LH.Result['configSettings']} settings
    * @return {{deviceEmulation: string, networkThrottling: string, cpuThrottling: string}}
    */
   static getEmulationDescriptions(settings) {
