+++ report/tsconfig.json
@@ -1,0 +1,39 @@
+{
+  "compilerOptions": {
+    "composite": true,
+    "outDir": "../.tmp/tsbuildinfo/report",
+    "emitDeclarationOnly": true,
+    "declarationMap": true,
+
+    // Limit to base JS and DOM defs.
+    "lib": ["es2020", "dom", "dom.iterable"],
+    // Include `@types/node` for use of `Buffer` in text-encoding.js.
+    "types": ["node"],
+    "target": "es2020",
+    "module": "es2020",
+    "moduleResolution": "node",
+
+    "allowJs": true,
+    "checkJs": true,
+    "strict": true,
+    // TODO: remove the next line to be fully `strict`.
+    "useUnknownInCatchVariables": false,
+
+    // "listFiles": true,
+    // "noErrorTruncation": true,
+    "extendedDiagnostics": true,
+  },
+  "include": [
+    "**/*.js",
+    "types/**/*.d.ts",
+  ],
+  "references": [
+    {"path": "../types/lhr/"},
+    {"path": "./generator/"}
+  ],
+  "exclude": [
+    "generator/**/*.js",
+    // These test files require further changes before they can be type checked.
+    "test/**/*.js",
+  ],
+}
