+++ tsconfig.json
@@ -20,7 +20,6 @@
     "root.js",
     "lighthouse-cli/**/*.js",
     "lighthouse-core/**/*.js",
-    "report/**/*.js",
     "clients/**/*.js",
     "build/**/*.js",
     "./types/**/*.d.ts",
@@ -39,7 +38,8 @@
   ],
   "references": [
     {"path": "./types/lhr/"},
-    {"path": "./report/generator/"}
+    {"path": "./report/"},
+    {"path": "./report/generator/"},
   ],
   "exclude": [
     "lighthouse-core/test/audits/**/*.js",
@@ -50,7 +50,6 @@
     "lighthouse-core/scripts/legacy-javascript/variants",
     "lighthouse-core/test/chromium-web-tests/webtests",
     // These test files require further changes before they can be type checked.
-    "report/test/**/*.js",
     "lighthouse-core/test/config/budget-test.js",
     "lighthouse-core/test/config/config-helpers-test.js",
     "lighthouse-core/test/config/config-plugin-test.js",
