+++ types/global-lh.d.ts
@@ -18,7 +18,6 @@ import LHError = require('../lighthouse-core/lib/lh-error.js');
 import LHResult from './lhr/lhr';
 import FlowResult_ from './lhr/flow';
 import Protocol_ from './protocol';
-import ReportResult_ from './html-renderer';
 import * as Settings from './lhr/settings';
 import StructuredData_ from './structured-data';
 import Treemap_ from './lhr/treemap';
@@ -63,7 +62,6 @@ declare global {
     export import FormattedIcu = I18n.FormattedIcu;
 
     export import Protocol = Protocol_;
-    export import ReportResult = ReportResult_;
 
     // lhr/settings.d.ts
     export import Locale = Settings.Locale;
